requirements.txt
Django>=1.9,<1.10
Markdown
celery>=3.1,<4.0
configobj
django-annoying
django-assets
django-async-messages
django-bitfield
django-bootstrap-form
django-braces
django-celery
django-csp
django-extensions
django-formtools
django-mptt
django-otp
django-ratelimit-backend
django-sendfile
django-session-csrf
django-two-factor-auth>=1.0.0
django-watson>=1.2.0
factory-boy
jsmin
lxml
markdown-newtab
premailer
progress
python-daemon<1.7
pytz
salmon-mail==2
simplejson
